requirements.txt
Package            Version
------------------ ---------
argh               0.26.2
astroid            2.3.3
bleach             3.2.1
Bootstrap-Flask    1.5.1
cachelib           0.1.1
certifi            2020.11.8
chardet            3.0.4
click              7.1.2
colorama           0.4.3
dnspython          1.16.0
email-validator    1.1.2
enum34             1.1.10
eventlet           0.29.1
Faker              5.0.0
Flask              1.1.2
Flask-Login        0.5.0
Flask-Moment       0.10.0
Flask-OAuthlib     0.9.6
Flask-SocketIO     4.3.2
Flask-SQLAlchemy   2.4.4
Flask-WTF          0.14.3
funcsigs           1.0.2
greenlet           0.4.17
html5lib           1.1
idna               2.10
importlib-metadata 3.1.1
ipaddress          1.0.23
isort              4.3.21
itsdangerous       1.1.0
Jinja2             2.11.2
lazy-object-proxy  1.4.3
Markdown           3.3.3
MarkupSafe         1.1.1
mccabe             0.6.1
mock               4.0.2
numpy              1.19.4
oauthlib           2.1.0
packaging          20.7
pathtools          0.1.2
pbr                5.5.1
Pillow             8.0.1
pip                20.2.4
Pygments           2.7.2
pylint             2.4.4
pyparsing          2.4.7
python-dateutil    2.8.1
python-dotenv      0.15.0
python-engineio    3.14.2
python-socketio    4.6.1
PyYAML             5.3.1
requests           2.25.0
requests-oauthlib  1.1.0
setuptools         40.8.0
six                1.14.0
SQLAlchemy         1.3.20
text-unidecode     1.3
typed-ast          1.4.1
urllib3            1.26.2
watchdog           0.10.4
webencodings       0.5.1
Werkzeug           1.0.1
wrapt              1.11.2
WTForms            2.3.3
zipp               3.4.0